+++ includes/class-activitypub.php
@@ -45,6 +45,10 @@ class Activitypub {
 
 		\add_action( 'in_plugin_update_message-' . ACTIVITYPUB_PLUGIN_BASENAME, array( self::class, 'plugin_update_message' ) );
 
+		\add_filter( 'activitypub_get_actor_extra_fields', array( self::class, 'default_actor_extra_fields' ), 10, 2 );
+
+		\add_action( 'tool_box', array( self::class, 'tool_box' ) );
+
 		// register several post_types
 		self::register_post_types();
 	}
@@ -130,11 +134,65 @@ class Activitypub {
 	}
 
 	/**
+	 * Add the 'self' link to the header.
+	 *
+	 * @see
+	 *
+	 * @return void
+	 */
+	public static function add_headers() {
+		// phpcs:ignore
+		$request_uri = $_SERVER['REQUEST_URI'];
+
+		if ( ! $request_uri ) {
+			return;
+		}
+
+		// only add self link to author pages...
+		if ( is_author() ) {
+			if ( is_user_disabled( get_queried_object_id() ) ) {
+				return;
+			}
+		} elseif ( is_singular() ) { // or posts/pages/custom-post-types...
+			if ( ! \post_type_supports( \get_post_type(), 'activitypub' ) ) {
+				return;
+			}
+		} else { // otherwise return
+			return;
+		}
+
+		// add self link to html and http header
+		$host      = wp_parse_url( home_url() );
+		$self_link = esc_url(
+			apply_filters(
+				'self_link',
+				set_url_scheme(
+					// phpcs:ignore
+					'http://' . $host['host'] . wp_unslash( $request_uri )
+				)
+			)
+		);
+
+		if ( ! headers_sent() ) {
+			header( 'Link: <' . $self_link . '>; rel="alternate"; type="application/activity+json"' );
+		}
+
+		add_action(
+			'wp_head',
+			function () use ( $self_link ) {
+				echo PHP_EOL . '<link rel="alternate" type="application/activity+json" href="' . esc_url( $self_link ) . '" />' . PHP_EOL;
+			}
+		);
+	}
+
+	/**
 	 * Custom redirects for ActivityPub requests.
 	 *
 	 * @return void
 	 */
 	public static function template_redirect() {
+		self::add_headers();
+
 		$comment_id = get_query_var( 'c', null );
 
 		// check if it seems to be a comment
@@ -311,36 +369,22 @@ class Activitypub {
 	}
 
 	/**
+	 * Adds metabox on wp-admin/tools.php
+	 *
+	 * @return void
+	 */
+	public static function tool_box() {
+		if ( \current_user_can( 'edit_posts' ) ) {
+			\load_template( ACTIVITYPUB_PLUGIN_DIR . 'templates/toolbox.php' );
+		}
+	}
+
+	/**
 	 * Theme compatibility stuff
 	 *
 	 * @return void
 	 */
 	public static function theme_compat() {
-		$site_icon = get_theme_support( 'custom-logo' );
-
-		if ( ! $site_icon ) {
-			// custom logo support
-			add_theme_support(
-				'custom-logo',
-				array(
-					'height' => 80,
-					'width'  => 80,
-				)
-			);
-		}
-
-		$custom_header = get_theme_support( 'custom-header' );
-
-		if ( ! $custom_header ) {
-			// This theme supports a custom header
-			$custom_header_args = array(
-				'width'       => 1250,
-				'height'      => 600,
-				'header-text' => true,
-			);
-			add_theme_support( 'custom-header', $custom_header_args );
-		}
-
 		// We assume that you want to use Post-Formats when enabling the setting
 		if ( 'wordpress-post-format' === \get_option( 'activitypub_object_type', ACTIVITYPUB_DEFAULT_OBJECT_TYPE ) ) {
 			if ( ! get_theme_support( 'post-formats' ) ) {
@@ -458,13 +502,43 @@ class Activitypub {
 			)
 		);
 
+		\register_post_type(
+			'ap_extrafield',
+			array(
+				'labels'           => array(
+					'name'          => _x( 'Extra fields', 'post_type plural name', 'activitypub' ),
+					'singular_name' => _x( 'Extra field', 'post_type single name', 'activitypub' ),
+					'add_new'       => __( 'Add new', 'activitypub' ),
+					'add_new_item'  => __( 'Add new extra field', 'activitypub' ),
+					'new_item'      => __( 'New extra field', 'activitypub' ),
+					'edit_item'     => __( 'Edit extra field', 'activitypub' ),
+					'view_item'     => __( 'View extra field', 'activitypub' ),
+					'all_items'     => __( 'All extra fields', 'activitypub' ),
+				),
+				'public'              => false,
+				'hierarchical'        => false,
+				'query_var'           => false,
+				'has_archive'         => false,
+				'publicly_queryable'  => false,
+				'show_in_menu'        => false,
+				'delete_with_user'    => true,
+				'can_export'          => true,
+				'exclude_from_search' => true,
+				'show_in_rest'        => true,
+				'map_meta_cap'        => true,
+				'show_ui'             => true,
+				'supports'            => array( 'title', 'editor' ),
+			)
+		);
+
 		\do_action( 'activitypub_after_register_post_type' );
 	}
 
 	/**
-	 * Add the 'activitypub' query variable so WordPress won't mangle it.
+	 * Add the 'activitypub' capability to users who can publish posts.
 	 *
 	 * @param int   $user_id  User ID.
+	 *
 	 * @param array $userdata The raw array of data passed to wp_insert_user().
 	 */
 	public static function user_register( $user_id ) {
@@ -472,5 +546,58 @@ class Activitypub {
 			$user = \get_user_by( 'id', $user_id );
 			$user->add_cap( 'activitypub' );
 		}
+	}
+
+	/**
+	 * Add default extra fields to an actor.
+	 *
+	 * @param array $extra_fields The extra fields.
+	 * @param int   $user_id      The User-ID.
+	 *
+	 * @return array The extra fields.
+	 */
+	public static function default_actor_extra_fields( $extra_fields, $user_id ) {
+		if ( $extra_fields || ! $user_id ) {
+			return $extra_fields;
+		}
+
+		$already_migrated = \get_user_meta( $user_id, 'activitypub_default_extra_fields', true );
+
+		if ( $already_migrated ) {
+			return $extra_fields;
+		}
+
+		$defaults = array(
+			\__( 'Blog', 'activitypub' )     => \home_url( '/' ),
+			\__( 'Profile', 'activitypub' )  => \get_author_posts_url( $user_id ),
+			\__( 'Homepage', 'activitypub' ) => \get_the_author_meta( 'user_url', $user_id ),
+		);
+
+		foreach ( $defaults as $title => $url ) {
+			if ( ! $url ) {
+				continue;
+			}
+
+			$extra_field = array(
+				'post_type'      => 'ap_extrafield',
+				'post_title'     => $title,
+				'post_status'    => 'publish',
+				'post_author'    => $user_id,
+				'post_content'   => sprintf(
+					'<!-- wp:paragraph --><p><a rel="me" title="%s" target="_blank" href="%s">%s</a></p><!-- /wp:paragraph -->',
+					\esc_attr( $url ),
+					$url,
+					\wp_parse_url( $url, \PHP_URL_HOST )
+				),
+				'comment_status' => 'closed',
+			);
+
+			$extra_field_id = wp_insert_post( $extra_field );
+			$extra_fields[] = get_post( $extra_field_id );
+		}
+
+		\update_user_meta( $user_id, 'activitypub_default_extra_fields', true );
+
+		return $extra_fields;
 	}
 }