+++ includes/class-scheduler.php
@@ -117,6 +117,11 @@ class Scheduler {
 	public static function schedule_post_activity( $new_status, $old_status, $post ) {
 		$post = get_post( $post );
 
+		if ( 'ap_extrafield' === $post->post_type ) {
+			self::schedule_profile_update( $post->post_author );
+			return;
+		}
+
 		// Do not send activities if post is password protected.
 		if ( \post_password_required( $post ) ) {
 			return;
@@ -130,9 +135,16 @@ class Scheduler {
 
 		$type = false;
 
-		if ( 'publish' === $new_status && 'publish' !== $old_status ) {
+		if (
+			'publish' === $new_status &&
+			'publish' !== $old_status
+		) {
 			$type = 'Create';
-		} elseif ( 'publish' === $new_status ) {
+		} elseif (
+			'publish' === $new_status ||
+			( 'draft' === $new_status &&
+			'draft' !== $old_status )
+		) {
 			$type = 'Update';
 		} elseif ( 'trash' === $new_status ) {
 			$type = 'Delete';
@@ -284,6 +296,7 @@ class Scheduler {
 		// the user meta fields that affect a profile.
 		$fields = array(
 			'activitypub_user_description',
+			'activitypub_header_image',
 			'description',
 			'user_url',
 			'display_name',
@@ -311,7 +324,9 @@ class Scheduler {
 
 	/**
 	 * Theme mods only have a dynamic filter so we fudge it like this.
-	 * @param  mixed $value
+	 *
+	 * @param mixed $value
+	 *
 	 * @return mixed
 	 */
 	public static function blog_user_update( $value = null ) {
@@ -321,6 +336,7 @@ class Scheduler {
 
 	/**
 	 * Send a profile update to all followers. Gets hooked into all relevant options/meta etc.
+	 *
 	 * @param int $user_id  The user ID to update (Could be 0 for Blog-User).
 	 */
 	public static function schedule_profile_update( $user_id ) {