+++ includes/model/class-blog.php
@@ -8,6 +8,7 @@ use Activitypub\Signature;
 use Activitypub\Activity\Actor;
 use Activitypub\Collection\Users;
 
+use function Activitypub\esc_hashtag;
 use function Activitypub\is_single_user;
 use function Activitypub\is_user_disabled;
 use function Activitypub\get_rest_url_by_path;
@@ -132,9 +133,15 @@ class Blog extends Actor {
 	 * @return string The User-Description.
 	 */
 	public function get_summary() {
+		$summary = \get_option( 'activitypub_blog_description', null );
+
+		if ( ! $summary ) {
+			$summary = \get_bloginfo( 'description' );
+		}
+
 		return \wpautop(
 			\wp_kses(
-				\get_bloginfo( 'description' ),
+				$summary,
 				'default'
 			)
 		);
@@ -182,7 +189,7 @@ class Blog extends Actor {
 	 * @return string The User-Name.
 	 */
 	public function get_preferred_username() {
-		$username = \get_option( 'activitypub_blog_user_identifier' );
+		$username = \get_option( 'activitypub_blog_identifier' );
 
 		if ( $username ) {
 			return $username;
@@ -230,11 +237,22 @@ class Blog extends Actor {
 	 *
 	 * @return array|null The User-Header-Image.
 	 */
-	public function get_header_image() {
-		if ( \has_header_image() ) {
+	public function get_image() {
+		$header_image = get_option( 'activitypub_header_image' );
+		$image_url    = null;
+
+		if ( $header_image ) {
+			$image_url = \wp_get_attachment_url( $header_image );
+		}
+
+		if ( ! $image_url && \has_header_image() ) {
+			$image_url = \get_header_image();
+		}
+
+		if ( $image_url ) {
 			return array(
 				'type' => 'Image',
-				'url'  => esc_url( \get_header_image() ),
+				'url'  => esc_url( $image_url ),
 			);
 		}
 
@@ -367,5 +385,68 @@ class Blog extends Actor {
 		} else {
 			return false;
 		}
+	}
+
+	/**
+	 * Get the User-Hashtags.
+	 *
+	 * @see https://docs.joinmastodon.org/spec/activitypub/#Hashtag
+	 *
+	 * @return array The User-Hashtags.
+	 */
+	public function get_tag() {
+		$hashtags = array();
+
+		$args = array(
+			'orderby' => 'count',
+			'order'   => 'DESC',
+			'number'  => 5,
+		);
+
+		$tags = get_tags( $args );
+
+		foreach ( $tags as $tag ) {
+			$hashtags[] = array(
+				'type' => 'Hashtag',
+				'href' => \get_tag_link( $tag->term_id ),
+				'name' => esc_hashtag( $tag->name ),
+			);
+		}
+
+		return $hashtags;
+	}
+
+	/**
+	 * Extend the User-Output with Attachments.
+	 *
+	 * @return array The extended User-Output.
+	 */
+	public function get_attachment() {
+		$array = array();
+
+		$array[] = array(
+			'type' => 'PropertyValue',
+			'name' => \__( 'Blog', 'activitypub' ),
+			'value' => \html_entity_decode(
+				sprintf(
+					'<a rel="me" title="%s" target="_blank" href="%s">%s</a>',
+					\esc_attr( \home_url( '/' ) ),
+					\esc_url( \home_url( '/' ) ),
+					\wp_parse_url( \home_url( '/' ), \PHP_URL_HOST )
+				),
+				\ENT_QUOTES,
+				'UTF-8'
+			),
+		);
+
+		// Add support for FEP-fb2a, for more information see FEDERATION.md
+		$array[] = array(
+			'type' => 'Link',
+			'name' => \__( 'Blog', 'activitypub' ),
+			'href' => \esc_url( \home_url( '/' ) ),
+			'rel'  => array( 'me' ),
+		);
+
+		return $array;
 	}
 }